Upgrade from `esbuild` `"^0.14.23"` to `"^0.17.06"`
