+++ lib/hanami/assets/compiler.rb
@@ -185,7 +185,7 @@ module Hanami
         result = destination_path
 
         if compile?
-          result.scan(/\A[[[:alnum:]][\-_]]*\.\w*/).first || result
+          result.scan(/\A[[[:alnum:]][-_]]*\.\w*/).first || result
         else
           result
         end
