+++ package.json
@@ -8,9 +8,10 @@
   "license": "MIT",
   "packageManager": "yarn@3.2.0",
   "devDependencies": {
-    "esbuild": "^0.14.23",
+    "esbuild": "^0.17.06",
     "esbuild-plugin-manifest": "^0.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
-  }
+  },
+  "dependencies": {}
 }
